Refactor opensearch lambda_handler
